supplier/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
